Move server/jest_configs/mapCoverage.js to a local npm package in jest-scripts/, as it will be common to multiple sub-repos but also has its own npm dependencies. Split mapCoverage.js in to a separate merge coverage reports and creat shield badge script
