+++ jest-scripts/.eslintrc.json
@@ -1,0 +1,39 @@
+{
+  "plugins": [
+    "node"
+  ],
+  "parserOptions": {
+    "ecmaFeatures": {
+      "experimentalObjectRestSpread": true
+    },
+    "ecmaVersion": 2017,
+    "sourceType": "module"
+  },
+  "extends": [ "eslint:recommended", "plugin:node/recommended" ],
+  "rules": {
+      "node/no-unsupported-features": ["error", {"ignores": ["modules"]}],
+      "indent": [
+          "warn",
+          2,
+          {
+            "SwitchCase" : 1,
+            "ignoredNodes": ["ConditionalExpression"],
+            "flatTernaryExpressions": true
+          }
+      ],
+      "comma-dangle":[ "warn", "always-multiline" ],
+      "linebreak-style": [
+          "off",
+          "unix"
+      ],
+      "semi": [
+          "off",
+          "always"
+      ],
+      "no-unused-vars": [
+        "warn",
+        { "args": "none" }
+      ],
+      "no-extra-semi": ["off"],
+  },
+}
