+++ jest-scripts/.gitignore
@@ -1,0 +1,8 @@
+*.DS_Store
+*.swp
+*.swo
+/node_modules
+.vscode/
+npm-debug*
+coverage
+*credentials*.json
